Fix test of split_graph_into_polylines()
