+++ BGL/test/BGL/CMakeLists.txt
@@ -35,6 +35,8 @@ if(OpenMesh_FOUND)
   target_link_libraries(graph_concept_OpenMesh PRIVATE ${OPENMESH_LIBRARIES})
 endif()
 
+create_single_source_cgal_program("test_split.cpp")
+
 create_single_source_cgal_program("next.cpp")
 
 create_single_source_cgal_program("test_circulator.cpp")
