--- BGL/test/BGL/split.cin
@@ -1,6 +1,0 @@
-5
-2 1 1 1
-1 1 1 0
-1 0 2 0
-2 1 2 0
-3 1 2 1
